fix: only import old localstorage in useOldLocalStorage
c'était pété la migration je crois @florianpanchout sans le stringify 🥲
